requirements.txt
accelerate==1.3.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
coloredlogs==15.0.1
einops==0.8.1
filelock==3.17.0
flatbuffers==25.1.24
fsspec==2025.2.0
huggingface-hub==0.28.1
humanfriendly==10.0
idna==3.10
Jinja2==3.1.5
MarkupSafe==3.0.2
moondream==0.0.6
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.2
onnx==1.17.0
onnxruntime==1.20.1
opencv-python==4.11.0.86
packaging==24.2
pillow==10.4.0
protobuf==5.29.3
psutil==6.1.1
pycparser==2.22
pyvips==2.2.3
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.2
setuptools==75.8.0
sympy==1.13.1
tokenizers==0.20.3
torch==2.6.0
tqdm==4.67.1
transformers==4.46.3
typing_extensions==4.12.2
urllib3==2.3.0
